requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2021.10.8
charset-normalizer==2.0.9
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.3
python-dotenv==1.0.1
python-telegram-bot==21.5
pytz==2024.1
requests==2.26.0
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
tzlocal==5.2
urllib3==1.26.7
wheel==0.44.0
